Added an hashtable implementation
remember to credit murmur3 from https://github.com/PeterScott/murmur3/ even though it's in the public domain :)

+++ syncqueue.c
@@ -101,7 +101,7 @@ void *syncqueuePop(SyncQueue *queue)
 }
 
 
-int syncqueueLen(SyncQueue queue)
+long syncqueueLen(SyncQueue queue)
 {
     return queue._len;
 }
